requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==22.2.0
chardet==4.0.0
charset-normalizer==3.1.0
discord.py==1.7.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
typing_extensions==4.5.0
yarl==1.8.2
